pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atarus-recon"
version = "0.5.0"
description = "External attack surface reconnaissance tool by Atarus Offensive Security"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Atarus Offensive Security", email = "info@atarussecurity.com"},
]
keywords = ["recon", "reconnaissance", "pentest", "security", "attack-surface"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
]
dependencies = [
    "click>=8.0",
    "dnspython>=2.4",
    "jinja2>=3.1",
    "rich>=13.0",
    "weasyprint>=60.0",
    "python-whois>=0.9",
    "requests>=2.31",
    "cryptography>=41.0",
]

[project.scripts]
atarus-recon = "atarus_recon.cli:main"

[project.urls]
Homepage = "https://github.com/atarus-security/atarus-recon"
Repository = "https://github.com/atarus-security/atarus-recon"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
atarus_recon = ["templates/*.html"]
